refactor: ♻️ move project, agent, preview, and github components to domain directories (Closes #220)
