+++ frontend/src/components/agent/AgentOutputViewer.tsx
