+++ frontend/src/components/agent/AgentPanel.test.tsx
@@ -2,18 +2,18 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import * as agentSessionsApi from '../api/generated/endpoints/agent-sessions/agent-sessions';
-import { useAgentStore } from '../stores/agentStore';
+import * as agentSessionsApi from '@/api/generated/endpoints/agent-sessions/agent-sessions';
+import { useAgentStore } from '@/stores/agentStore';
 import { AgentPanel } from './AgentPanel';
 
-vi.mock('../api/generated/endpoints/agent-sessions/agent-sessions', () => ({
+vi.mock('@/api/generated/endpoints/agent-sessions/agent-sessions', () => ({
   useListAgentSessions: vi.fn(),
   useStartAgentSession: vi.fn(),
   useStopAgentSession: vi.fn(),
   useGetAgentSessionOutputs: vi.fn(),
 }));
 
-vi.mock('../hooks/useAgentEvents', () => ({
+vi.mock('@/hooks/useAgentEvents', () => ({
   useAgentEvents: vi.fn(),
 }));
