+++ frontend/src/components/preview/PreviewPanel.test.tsx
@@ -2,11 +2,11 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import * as previewsApi from '../api/generated/endpoints/previews/previews';
-import type { DockerPreview } from '../api/generated/model';
+import * as previewsApi from '@/api/generated/endpoints/previews/previews';
+import type { DockerPreview } from '@/api/generated/model';
 import { PreviewPanel } from './PreviewPanel';
 
-vi.mock('../api/generated/endpoints/previews/previews', () => ({
+vi.mock('@/api/generated/endpoints/previews/previews', () => ({
   useListPreviews: vi.fn(),
   useCreatePreview: vi.fn(),
   useDeletePreview: vi.fn(),
