+++ frontend/src/components/project/ProjectCreateDialog.test.tsx
@@ -2,11 +2,11 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { act, render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import * as projectsApi from '../api/generated/endpoints/projects/projects';
-import { useUiStore } from '../stores/uiStore';
+import * as projectsApi from '@/api/generated/endpoints/projects/projects';
+import { useUiStore } from '@/stores/uiStore';
 import { ProjectCreateDialog } from './ProjectCreateDialog';
 
-vi.mock('../api/generated/endpoints/projects/projects', () => ({
+vi.mock('@/api/generated/endpoints/projects/projects', () => ({
   useCreateProject: vi.fn(),
 }));
