+++ frontend/src/components/project/ProjectSettingsModal.test.tsx
@@ -2,15 +2,15 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import * as membersApi from '../api/generated/endpoints/project-members/project-members';
-import * as projectsApi from '../api/generated/endpoints/projects/projects';
-import type { Project, ProjectMember } from '../api/generated/model';
-import { MemberRole } from '../api/generated/model';
-import { useAuthStore } from '../stores/authStore';
-import { useUiStore } from '../stores/uiStore';
+import * as membersApi from '@/api/generated/endpoints/project-members/project-members';
+import * as projectsApi from '@/api/generated/endpoints/projects/projects';
+import type { Project, ProjectMember } from '@/api/generated/model';
+import { MemberRole } from '@/api/generated/model';
+import { useAuthStore } from '@/stores/authStore';
+import { useUiStore } from '@/stores/uiStore';
 import { ProjectSettingsModal } from './ProjectSettingsModal';
 
-vi.mock('../api/generated/endpoints/projects/projects', () => ({
+vi.mock('@/api/generated/endpoints/projects/projects', () => ({
   useGetProject: vi.fn(),
   useUpdateProject: vi.fn(),
   useDeleteProject: vi.fn(),
@@ -18,7 +18,7 @@ vi.mock('../api/generated/endpoints/projects/projects', () => ({
   getGetProjectQueryKey: vi.fn((id: string) => ['/api/projects', id]),
 }));
 
-vi.mock('../api/generated/endpoints/project-members/project-members', () => ({
+vi.mock('@/api/generated/endpoints/project-members/project-members', () => ({
   useListMembers: vi.fn(),
 }));
